feat: change files names to EmojiReactions and refactor references
